cr2 is 64-bit
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@28 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/arch/amd64/x86.h
@@ -22,7 +22,7 @@ static __inline void lldt(uint16_t sel) __attribute__((always_inline));
 static __inline void ltr(uint16_t sel) __attribute__((always_inline));
 static __inline void lcr0(uint32_t val) __attribute__((always_inline));
 static __inline uint32_t rcr0(void) __attribute__((always_inline));
-static __inline uint32_t rcr2(void) __attribute__((always_inline));
+static __inline uint64_t rcr2(void) __attribute__((always_inline));
 static __inline void lcr3(uint64_t val) __attribute__((always_inline));
 static __inline uint64_t rcr3(void) __attribute__((always_inline));
 static __inline void lcr4(uint32_t val) __attribute__((always_inline));
